Let `CallGraph.HigherOrderCallGraph` contain a `MutableDefineCallGraph`
Summary:
Currently, `CallGraph.HigherOrderCallGraph` contains a `DefineCallGraph`, which
is challenging to merge with another `DefineCallGraph`. For example, it is
unclear how to merge a `Singleton` with a `Compound`, because a `Singleton`
does not contain the expression identifiers that are required in a `Compound`.
Merging two call graphs will be useful later for defining the `join` operation
in `Model` for the global fixpoint algorithm.

This diff lets `CallGraph.HigherOrderCallGraph` contain a
`MutableDefineCallGraph`, which can be easily merged.

Note that merging two call graphs is not going to be used in the near future,
since `Model.join` is only used in analyzing overrides and we plan to skip
analyzing overrides for now. However this will eventually be useful when
analyzing overrides.

Reviewed By: arthaud

Differential Revision: D65562828

fbshipit-source-id: c719877ec060652a31920b3b24425f46dbf3cf84