Make constraints explicit in `solve_less_or_equal`
Summary:
Same as the previous commit, this commit is me trying to make the existing
implementation more usable as a quick guide to how typing features relate to constraint
solving by making it implicit in the huge match statements of the solver that we usually
just propagate the initial constraints (in other words, constraints are just context for
most of the operations), and making it explicit when the constraints being passed might
actually be modified.

This helps me see which type system features are most tied in with constraint
solving with an easy search instead of having to read line-by-line (before, we
were often shadowing the initial constraints with a variable named
`constraints` so figuring this out was highly nonlocal - it basically required
reading every line of code and was entirely un-searchable).

This commit cuts over `solve_less_or_equal` to the new style.

Reviewed By: samwgoldman

Differential Revision: D65961791

fbshipit-source-id: 9e4f76d5de50018e6d41862afd6712166c808068