Make constraints explicit in `solve_ordered_types_less_or_equal`
Summary:
This commit continues my work making it so that by default, constraints are passed
around without modification in the constraint solver and it is explicit when we might
modify the original constraints of a top-level function.

Since only a relative handful of operations actually affect constraints, this will make
it much easier to quickly search for parts of the type system that actually interact
nontrivially with constraints, which could be helpful particularly when using the
code as a reference.

This wraps up my stack, the remaining operations (mainly related to recursive types)
don't even take constraints as inputs.

Reviewed By: samwgoldman

Differential Revision: D65961790

fbshipit-source-id: e185a03ef9928b80ae696656f47d318c808ecba7