Tweak cycleDetection doc
Summary:
I did some thinking about this because it impacts any rewrite of the constraint
solver - callable resolution does indeed have a truly recursive edge case, where
the `__call__` attribute of a class is annotated as a class that itself is the
current class or creates any cycle back to the current class.

In practice it mostly happens because of our made-up BoundMethod type, but even
in the normal python type system it is possible to construct recursive examples.

Reviewed By: rchen152

Differential Revision: D65958795

fbshipit-source-id: 666f908c95ea2d953bc49b461227972532a48733