Define different versions of `build_whole_program_call_graph` for `DefineCallGraph` and `MutableDefineCallGraph`
Summary:
For context, `DefineCallGraph` and `MutableDefineCallGraph` both define the
call graph of a callable, but the `mutable` version supports faster mutation
and stores additional expression identifiers, when multiple of them exist at
the same location.

Currently, `build_whole_program_call_graph` only builds call graphs in the type
of `DefineCallGraph`, but not `MutableDefineCallGraph`. It would be useful to
let the function generate both call graphs:
1. While unit-testing the original call graph building requires call graphs in
the type of `DefineCallGraph`, unit-testing higher order call graph building
requires call graphs in the type of `MutableDefineCallGraph`.
2. We may want to keep the higher order call graph building optional (e.g., in
case it is too slow) for the taint analysis, in which case we want to build
call graphs in the types of both `DefineCallGraph` and
`MutableDefineCallGraph`.

Reviewed By: arthaud

Differential Revision: D65785112

fbshipit-source-id: 28bc6914c8320746028b55f1cfacf3b44c5e3c2c