Update `MutableDefineCallGraph`’s type into a serializable map
Summary:
So that the call graphs of this type
1. can be stored into the shared memory (so that they can be queried during the
taint analysis).
2. can be merged properly, which will be useful for later to merge two
`MutableDefineCallGraph` in the global fixpoint computation.

Note that with this change, it could be slower to update the map than using a
hash map. But we cannot use a hash map because it cannot be stored in the
shared memory.

Reviewed By: arthaud

Differential Revision: D65787084

fbshipit-source-id: a173d459aaa4668aed6111f107b29eddb66e3385