Scaffold a global fixpoint computation for higher order call graph building
Summary:
We have to create a new file to define the global fixpoint, instead of defining
it inside `callGraph.ml`, due to avoiding circular dependencies:
1. `callGraph.ml` depends on `fixpointAnalysis.ml`
2. `fixpointAnalysis.ml` depends on `dependencyGraph.ml`
3. `dependencyGraph.ml` depends on `callGraph.ml`

Note that this diff does *not* attempt to correctly implement the global
fixpoint algorithm, which will be done in the follow-ups. Instead, this diff
aims at defining the "skeleton" of the global fixpoint algorithm and setting up
the tests.

Reviewed By: arthaud

Differential Revision: D65562827

fbshipit-source-id: 5aabf014fdad7b3309403b2406fda5357107a977