Allow global fixpoint algorithm to use additional dependencies and analyze newly found callables
Summary:
So that the dependency graph can be updated with the additional dependencies
after each iteration. For context, the current global fixpoint algorithm takes
a dependency graph as inputs and does not allow the graph to be changed by any
iteration.  To compute the higher order call graphs, each iteration could
produce new dependencies.

In addition, the new dependencies might be new callables that are not included
in the initially discovered callables. Hence, we need to analyze the newly
discovered callables in the next iteration.

Reviewed By: arthaud

Differential Revision: D65555105

fbshipit-source-id: 5ae7f31c7ab8c88b4546b6889f16cc2a56ea5bcf