Add and use --write-acutal-results-on-failure for pysa tests
Summary:
The `results.json` used in github CI needs to match `results.actual`,
and frequently the results change when Pysa improves and we forget to
update the tests.

It would be nice if it were very easy to update tests based on red
github CI; dumping the results should make this much easier, the diff
views are unusable for actually fixing tests in most cases.

Differential Revision: D66269136