Simplify literal code: push get-enum-class logic to `classes.rs`
Summary:
The `Lit` code has too low-level of coupling to class logic.

It already is taking an identifier-to-class callback and an mro callback,
and it still has to blindly produce potentially invariant-violating `ClassType`
values.

Instead, let's move this logic into a helper in `classes.rs` that can much
more reasonably handle all the things we need to do here.

Reviewed By: ndmitchell

Differential Revision: D66341306

fbshipit-source-id: c34084bffd5779c9ac84bbe737bbf63d236f698f