Make `ClassType` struct internals private
Summary:
I am working on fixing legacy generic parameter handling,
and I ran into a situation where I'm seeing `targs` and `tparams`
not aligned.

This points at a smell in the code: by having `ClassType` as a public
struct, we make it far too easy to blindly create instances, which is
known to panic Pyre later if the `targs` were invalid.

Instead, we ought to make the struct private and use a constructor
to make the invariant clearer. Ideally the constructor would validate
the invariant, but this runs into a couple of practical issues: the
actual invariant is checked inside of `AnswersSolver` methods that need
access to all the calculations, and we need an unsafe version for now
anyway for `stdlib` in its curent form.

That's what this commit does: we have a `create_with_validated_targs`
to create `ClassType`, and the name is supposed to suggest even to someone
new to the codebase that there are invariants in play here.

It also makes tracking down `ClassType` creation much easier if we do
run into panics, because now we know that bad creates *have* to be
triggered from some `create_with_validated_targs` call.

Reviewed By: ndmitchell

Differential Revision: D66341308

fbshipit-source-id: aac3fcd87976eb6632f75dbca017198d630263de