Remove Self type handling from base classes
Summary:
There are dual motivations for this change:

The first is that recursive bindings are hard, we have a panic-inducing bug
right now where *if* we start using the bases to set class `tparams`, we wind
up hitting recursion, creating empty `tparams` inside the recursion, then
instantiating a malformed `self` type that panics the type checker later.

That *could* be handled via indirection, by making two kinds of base class
representations, one before handling self type that is used for `tparams` and
the other after handling self type.

But that comes to the second motivation: the spec clearly states that `Self`
is disallowed in base classes anyway. So we can just use the simpler solution
of not handling it at all:
https://github.com/python/typing/blob/main/conformance/tests/generics_self_usage.py#L100-L103

Note that eventually we're going to need to make it an error if any type
is `Self` after some stage of evaluation (this isn't really any different
from type var declarations - in both cases the name has to be in static analysis
scope); I'm not worrying about that logic yet since we need it not just on base
classes but *everywhere* that `Self` isn't scoped, and it's not a blocker on
the kind of basic language support we are trying to hurry along right now.

Also note that our implementation of `Self` needs to be fully rewritten
eventually, since it is supposed to behave as a type variable with special
semantics whereas we currently approximate it as a macro that expands to the
current class type. This is another reason I don't want to worry about errors
yet, too much is going to change for it to be worth investing right now.

Reviewed By: ndmitchell

Differential Revision: D66384568

fbshipit-source-id: e288c4ffe94c8b631fa333d35b7d0b2316a1eadb