Fix legacy generic tparam ordering
Summary:
We realized that Pyre is not handling legacy generics correctly, it's
creating legacy type parameters in the order they appear syntactically
in base classes.

What we actually want is more complicated, we (probably) want
`[*scoped_tvars, *generic_tvars, *protocol_tvars, *implicit_tvars]`
where only the `implicit_tvars` are coming in syntactic order. I
say 'probably' because code that uses all of these is ill-typed, but
we still want well-defined behavior and I think this ordering is
the best choice.

In particular, on a prototypical legacy class like
```
class Foo(Baz[S], Bar[T], Generic[T, S]):
    pass
```
we want the type parameters to be `T, S`, and the current logic is giving
`S, T` which causes issues with typeshed stubs.

This stack will fix the problem. This diff refactors and simplifies a bit in a
way that will make it easier for me to stack the actual fix, which will require
traversing the bases.

Reviewed By: ndmitchell

Differential Revision: D66384567

fbshipit-source-id: acb24b420627b2071443e249d3a81f205d3f1058