Move class def handling to `classes.rs`
Summary:
It really always should have been here, but we weren't actually
doing much so it didn't matter. In an upcoming refactor, the logic
will be more involved because we'll do the work needed to extract
type parameters while evaluating the class definition, at which point
it becomes more important for this to live in the `classes.rs` module
so that it can use other private class-related helpers.

Reviewed By: ndmitchell

Differential Revision: D66526169

fbshipit-source-id: 2ecb32cf8802ebff01ea393765a6193eec3406fd