Make LookupAnswer a trait
Summary:
Required to allow it to be a function. Alas the trait isn't object safe, so I have to dump <Ans> in a million places and split the Solver trait. All very annoying, but I don't see a way to avoid it and make Answers a function, which is required to do any smart scheduling.

Will probably rethink the Table abstraction over thanksgiving to see if we can do better.

Reviewed By: grievejia

Differential Revision: D66543501

fbshipit-source-id: f69d6127284486bfba035562be846b7ba34edbb0