Write into a separate `DefineCallGraph` than the input `DefineCallGraph` when building higher order call graphs
Summary:
Otherwise, the same `DefineCallGraph` is being both read from and written into.
This is not good practice and also problematic because it is possible to write
new targets (e.g., parameterized targets) for an `Expression.Call`, which will
then be read upon the next visit to the expression as the targets from the
original `DefineCallGraph`.

Reviewed By: arthaud

Differential Revision: D66503635

fbshipit-source-id: cec1c5cb50b31258ad8d7181dbb219361daad09d