Enforce the `targs` vs `tparams` invariant on `ClassType` creation :)
Summary:
We've had several bugs caught in the wild (as in, not by unit tests) tied to
edge cases where a `ClassType` could be created with invalid type arguments;
often this was due to nonobvious recursive behaviors in `AnswersSolver`.

We really want to be able to rule this possibility out entirely,
which we can do if the constructors enforce that `targs` and `tparams` line up.

This commit makes that change: instead of checking the invariant when we create
a `substitution`, we check it on class creation.

Assuming there are no bugs in `pyre2`, we should never hit this in user code. One
challenge, however, is that `stdlib.rs` makes assumptions about the stdlib stubs
that we cannot intrinsically enforce. It's hard to know how to handle this - the
type checker is broken if the stdlib stubs don't match assumptions - but for now
panicking makes sense because only our team works with pyre2, and we should go
fix the problem. I do think it's helpful to customize the error in that case so
that it's very clear what is causing the panic.

Reviewed By: ndmitchell

Differential Revision: D66546444

fbshipit-source-id: 68ab2e8966ea368c40691520c33d488fc1511c29