Ensure pseudo method calls evaluate arguments (1/3)
Summary:
We use the "call_method" helper to evaluate various pseudo method calls. However, we only performed call inference when the method was found and actually callable.

This behavior leads to missed errors, both when the receiver of the call itself is `Any` (fixed in this diff) and when the receiver can not be computed, resulting in an error (see next diff).

This behavior does not affect actual method calls, which go through a different code path.

Reviewed By: ndmitchell

Differential Revision: D66563272

fbshipit-source-id: 4fa537dc431068c1c847995f71786237deba619d