Move keywords into the `Mro` binding
Summary:
After considering options, I think it makes sense to handle class
keywords in the same "metadata" calculation as the MRO - my rationale
is that:
- we're going to need all of the keywords in a map somewhere eventually,
  because type checking the class definition will eventually require
	validating the keywords
- given that we need the map, it's nice to put it with the MRO to reduce
  the total number of bindings needed... the code is simpler and we probably
	save some RAM by not creating tons of bindings.
- a potential counterargument that would nudge us toward having more bindings
	would be that it's better to be as lazy as possible, but in most cases we're
	going to need keywords at the same time as the MRO anyway, because both of
	them will ultimately wind up needing to get resolved before we can properly
	analyze attribute access (attributes depend on class keywords through
	metaclass as well as special-case typing features like typed dicts and
	dataclasses).

This diff does the work for bindings. At this stage I still do not actually
store the results of *evaluating* keywords anywhere, but we now perform
the type check as part of evaluating the Mro binding.

NOTE: the name `Mro` isn't very good once we start moving more things in.
But I'd rather defer the rename because I know there will be merge conflicts
if I try to rename right now; I put a TODO in to handle it later in this
stack.

Reviewed By: rchen152

Differential Revision: D66681754

fbshipit-source-id: 82b91aea323b040cdad8e0b956a7c38829c0f57c