Rename `Mro` to `ClassMetadata`
Summary:
I want to generalize the data type currently named `Mro` to
contain not only information about the "normal" ancestors, but also
class keywords and the metaclass.

My rationale for keeping these together is twofold:
- First, it is easier to maintain code that has a relatively minimal
  number of bindings for analyzing a concept.
	- Because of this principle, I want to avoid the class keywords
	  and the metaclass being a totally separate binding from the Mro;
		keeping them together is probably a little more space-efficient
		and also just makes the code easier to understand.
	- Note that we *do* really need `ClassMetadata` to be separate from
	  `ClassDef`, for both efficiency and correctness:
		- In terms of efficiency, we don't want to pay the cost of computing the
			Mro when we look up the class def, because there might be many cases
			where a class appears as a type but we never need to handle attribute
			access in which case we can skip the Mro analysis.
		- In terms of correctness, it's very hard to handle the class metadata
		  (in which the class itself might appear as a type argument to base
			classes or keywords) without recursion bugs.
- Second, there's *not* much extra compute cost, because to handle
  all cases correctly (including typed dicts, base classes, and custom
	metaclasses) we're going to have to analyze the keywords any time we
	deal with attribute access, which means the keywords and the Mro are
	usually needed at the same time. In addition, metaclass and the Mro have the
	same fanout to all ancestors, so they really do make a lot of sense to
	compute together.

Obviously `Mro` is a bad name for this concept as soon as we have more than
just ancestors in the data structure, so I'm renaming it. As a bonus, my IDE
will stop trying to open the `mro` tests every time I type `command-P mro.rs`
which I view as a nice bonus :)

----

I'm not sure `ClassMetadata` is the best name for the concept of "type
information derived from the class header", which is more or less what I want
this binding to represent, but it's the best I can come up with... Pyre1
already uses this term for similar information so it's already familiar, and it
has the advantage of not being too specific (I considered `Inheritance`, but
that sounds too much like the same thing as the `Mro`, which is going to wind
up being only one part of the new datatype).

It's easy to rename if we think of something better later, for now I'm just
going with something better than `Mro`.

Reviewed By: rchen152

Differential Revision: D66681753

fbshipit-source-id: fd220efc05fa394f4d223d89ac3d78a1ec9cddbb