support functional syntax for StrEnum and IntEnum
Summary:
Context: https://github.com/facebook/pyre-check/issues/946

Previously, Pyre only supported the functional declaration syntax for `enum.Enum` and not any of its subclasses. It's difficult to extend this to all subclasses due to this being handled in preprocessing, but this diff extends its support to `enum.StrEnum` and `enum.IntEnum` which is a fairly easy lift.

Reviewed By: stroxler

Differential Revision: D66669647

fbshipit-source-id: b4560e637a1806e195d21be7adcdf0e06ed4d016