Do not visit try statements during the call graph building
Summary:
Since D58018874, the local fixpoint explicitly visit the `Statement.Try`, although that statement is lowered down into multiple nodes. This basically mean that we visit statements inside the `try` twice: once when we see the `Cfg.Node.Try`, and once when we actually visit the `Cfg.Node.Block` containing the actual statements.

This leads to wrong sink indices for our handles. Let's fix it.

Reviewed By: tianhan0

Differential Revision: D66754610

fbshipit-source-id: 68a07dcd230ad6e36712be0cc4799ba1fc900e1a