simplify match logic and implement basic MatchAs
Summary:
This diff refactors the match logic to remove the `bind_target` call, which doesn't make sense since `MatchValue` only allows literals/constants.

If the thing being matched on is a name, we _could_ do a `bind_target` on that to refine the type, but I think it may be best to save all the narrowing for later.

The next diff will probably factor this into a recursive function.

Reviewed By: ndmitchell

Differential Revision: D66779115

fbshipit-source-id: ec4ed07415ad5cd7b4414790397d95f9ddd3c70f