Introduce Variable::Unwrap, fix unwrap_awaitable
Summary:
Using Contained vars causes the solver to promote literals to their general types, so `Literal[42]` becomes `int` when unwrapping.

For now, I am going to work around this by just creating a new type which does will not have the promotion behavior. This solving code is inevitably going to change as we get more into instantiation, so I don't want to overthink this right now.

Reviewed By: ndmitchell

Differential Revision: D66895963

fbshipit-source-id: 10e38645e1f8d49f792b2fa079c171f75b63c87b