implement MatchClass
Summary:
Specs:

https://docs.python.org/3/reference/compound_stmts.html#class-patterns

https://docs.python.org/3/library/ast.html#ast.MatchClass

This diff implements both positional and keyword matching for class patterns. For positionals, I look up the index in the `__match_args__` attr to get the name of the actual attr to look up.

Like the MatchMapping diff, I don't refine or check the actual value being matched is the right class. I'll do that all later.

The documentation specifies that duplicate keys will raise an error at runtime. I don't check against this right now, but neither do Mypy and Pyright. There may be an opportunity to add this to the spec.

Reviewed By: ndmitchell

Differential Revision: D66857624

fbshipit-source-id: 5cc885da6c60cd2164cf9fded10efd74efa6f371