Remove a TODO related to self-attr-assignment
Summary:
I added this TODO when I was working on some class logic because I at least
thought we were unconditionally defining attributes based on assignment
(as Pyright does), which we may well decide to do in the future but we'd
probably want it to be at least some kind of strict-mode error in order
to match Pyre1's level of initialization safety.

I'm not sure anymore whether the logic since changed or I misunderstood in the
first place, but regardless it is currently the case that we only actually
use attributes in `__init__` so this TODO is no longer accurate.

Reviewed By: rchen152

Differential Revision: D66998662

fbshipit-source-id: 599ff192ada992576bb546545b0d28ce423efb46