Remove github CI linting (#950)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/950

Most of the github signals are valueable to us: make sure that
the open-source build works on stock OCaml + rust, make sure that
the frontend runs properly on numerous python versions, etc.

The lint on the other hand is low-value. We have internal linters,
which have slightly different settings that can conflict with the
vanilla CI lints: applying open-source autoformat is always okay
internally but not vice versa.

As a result, I'd like to propose that we get rid of the open source
lint CI action - we only need one source of truth. I'm *not* getting
rid of the pre-commit integration because autoformatting hooks are
still useful to external developers - it's only the red diff signals
that are unhelpful.

Reviewed By: grievejia

Differential Revision: D67044600

fbshipit-source-id: ec69d5aedcdaab2e0cb173a900d7421f18c8bd87