Delay handling of sourcedb module name conflicts
Summary:
code_refactoring

Previously, we build the buck sourcedb by loading&parsing the manifest files and then immediately assume that the loaded manifest entries do not have conflict. If any of the source/dependency/typeshed manifest files contain multiple entries that correspond to the same module name, we naively resolve the conflict with the "last occurrence win" policy.

Such a scheme has very apparent drawbacks. For example, if the source target owns both `foo/bar.pyi` and `foo/bar.py`, then we'll randomly prioritize one of them for module `foo.bar` depending on which item appear first in the manifest. But the right thing to do would be to prioritize the `.pyi` over the `.py` file. On the other hand, if another file in the repo contains `import foo.bar` and want to go-to-def for `foo.bar`, there's also an argument that we should offer `foo/bar.py` as candidate.

The bottom line here is that different use cases may want different ways to resolve module name conflicts in sourcedb, and it's probably desirable to avoid committing to anything eagerly upfront as we did before. This diff makes it such that we no longer do "last occurrence win" immediately when reading the manifest files -- instead we preserve all entries including the conflicting ones, and delay the resolution a little bit later. We'll still eventually get the "last occurrence win" behavior in this diff (to make the diff smaller), but the setup now makes it possible to change that in D67068082.

Reviewed By: ndmitchell

Differential Revision: D67059758

fbshipit-source-id: 903d6084ad4691a363272801560ee51075e1a4a0