Create `ResolvedExpression` for calls to decorators
Summary:
The vision is that, we eventually turn ASTs into `ResolvedExpression`s, which
are ASTs with callee information. This is similar to typed ASTs. The only
difference is that we store callee information, instead of types.

For function `foo`, when it is decorated with function `decorator`, we create
AST `decorator(foo)` and call `visit_expression` on it to resolve callees. The
resultant `DefineCallGraph.t` will be used by the higher order call graph
building, to obtain the returned callables from `decorator(foo)`.
This can be realized by defining two versions of `analyze_expression`: One that
takes `Expression.t` as inputs, and another that takes `ResolvedExpression.t`
as inputs. We can further avoid code duplication by abstracting the
commonalities:
1. An AST node. This could be `Expression.t` or an artifically created
`Expression.t`.
2. How to find callees on any sub-tree within an AST node.

Reviewed By: arthaud

Differential Revision: D67003798

fbshipit-source-id: eac5acaf6603e48f330017c65750499f3f4a0da0