Add a time buffer for background logging on unknown failures
Summary:
I spent almost half a day debugging a github action failure this week whose
root cause was a glibc version mismatch (it is not the only time this has
happened - our releases are built on fairly new linux images, which can
very easily result in pinned CI systems trying to run them with incompatible
system library versions).

The painful thing is all that debugging was really just me getting to the
actual error at which point I knew exactly what to do - bump the image for the
CI action.

Why wasn't the error message shown? The problem is that our foreground thread
exits immediately after the backend finishes, but there's a background logging
thread that is essentially tailing the backend's stderr and that comes with
delays. When the foreground thread exits, the background thread shuts down
(which is what we *want* normally, e.g. in `pyre incremental` the backend keeps
running but we don't want to tail the logs, we want to exit when the frontend
finishes).

But in the case of sudden backend failures like a bad glibc version, it all
happens so fast that we lose the error messages.

The ideal way to fix this would be to restructure our entire system so that
the background logging and foreground communicate in two-way channels,
so that the foreground would signal to shut down, the background would finish
flushing logs and signal when it is ready, and then we would exit. But that
would take days for me to implement, particularly because I'm not very up
to date on python concurrency primitives. It works just as well in practice to
simply wait for a half second before exiting, that's a very long time compared
to typical log tailing latency.

Reviewed By: grievejia

Differential Revision: D67127220

fbshipit-source-id: e85c30c8560a54e7effdb0aa88057fde59a30a74