Make ClassAttributeIsGeneric Lookup::Error instead of Lookup::NotFound
Summary:
For code which needs to do lookup fallback behavior, it's not correct to fall back when accessing a generic class method.

Before this diff, generic class access was treated as a "not found" result. With this change, callers who care will be able to avoid incorrect fallback.

I considered alternatively returning a Found type with optional "warning." It may turn out that this is a better API, but for now this change is simpler.

Reviewed By: ndmitchell

Differential Revision: D67160890

fbshipit-source-id: 59a32a1571fc6e5a89e534b12773cea805bd6157